LayoutTests/imported/w3c/web-platform-tests/workers/opaque-origin-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: BroadcastChannel


PASS Worker has an opaque origin.
FAIL Worker can read its own blobs. assert_equals: expected (string) "from worker" but got (object) null
FAIL Worker can read its owners blobs. assert_equals: expected (string) "from page" but got (object) null
FAIL Worker can XHR fetch a blob. assert_equals: expected "from page" but got ""
FAIL Worker can fetch a blob. promise_test: Unhandled rejection with value: object "TypeError: Cross origin requests are only supported for HTTP."
FAIL Worker can access BroadcastChannel Can't find variable: BroadcastChannel
